--- a/nunchaku/caching/utils_v2.py
+++ b/nunchaku/caching/utils_v2.py
@@ -27,7 +27,7 @@ from typing import Any, Dict, Optional, Union
 import torch
 from diffusers.models.modeling_outputs import Transformer2DModelOutput
 
-from nunchaku.caching.FBCache import check_and_apply_cache
+from nunchaku.caching.fbcache import check_and_apply_cache
 from nunchaku.models.embeddings import pack_rotemb
 from nunchaku.models.transformers.utils import pad_tensor