@Bluear7878 2025-08-19
> Could you provide the reference cache implementation in diffusers?

I looked into implementing caching in diffusers.
FasterCache: [link](https://huggingface.co/docs/diffusers/main/en/optimization/cache)

FBCache (in ParaAttn): [link](https://huggingface.co/docs/diffusers/main/en/optimization/para_attn)

DeepCache: [link](https://huggingface.co/docs/diffusers/main/en/optimization/deepcache)

Should I align my current implementation with the style and structure of these existing caches?
Any specific one I should prioritize as a reference?